requirements.txt
azure-functions
azure-mgmt-resource
azure-mgmt-network
azure-mgmt-compute
azure-identity
msrestazure
